chore: refactor nav/menu
- extract config
- split mobile/desktop menus into separate components
- extract NavLink component
